requirements.txt
fastapi
uvicorn
pypdf
python-docx
python-dotenv
openai
chromadb